procedure_generator/requirements.txt
fillpdf==0.7.3
Gooey==1.0.8.1
PyMuPDF==1.26.1
pyinstaller==6.14.1
python-docx==1.2.0
six==1.17.0
playwright>=1.52.0
PyYAML>=6.0